requirements.txt
gunicorn==19.9.0
MarkupSafe==1.1.1
Flask==1.1.2	
Jinja2==2.11.2	
Pillow==7.2.0	
Werkzeug==1.0.1	
itsdangerous==1.1.0	
joblib==0.15.1	
kiwisolver==1.2.0	
matplotlib==3.2.1	
numpy==1.18.4	
opencv-python==4.2.0.34	
opt-einsum==3.2.1	
pandas==1.0.3	
pip==20.1.1	
requests==2.23.0	
scikit-learn==0.23.1	
scipy==1.4.1
seaborn==0.10.1
six==1.15.0	
sklearn==0.0
urllib3==1.25.9	

